Revised for SudachiPy v0.5.2 and update json files for holo-medel.
