+++ .gitignore
@@ -1,0 +1,26 @@
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+*$py.class
+
+# Distribution / packaging
+.Python
+env/
+build/
+develop-eggs/
+dist/
+eggs/
+.eggs/
+*.egg-info/
+
+# virtualenv
+.venv/
+venv/
+ENV/
+
+# no docs
+docs/
+
+# no sample
+sample/
+
