+++ models/holo/conv_auxverb.json
@@ -13,7 +13,7 @@
 		"う": "んじゃ",
 		"？": "かや？",
 		"。": "んじゃ。",
-		"": "んじゃ"
+		"EOS": "んじゃ"
 	},
 	"でしょう": {
 		"か,ね,？": "かのぉ？",
@@ -26,56 +26,65 @@
 		"ね": "じゃろ。",
 		"よ": "じゃよ。",
 		"。": "じゃ。",
-		"": "じゃ。",
-		"？": "じゃろ？"
+		"？": "じゃろ？",
+		"EOS": "じゃ。"
 	},
 	"でしょ": {
 		"？": "じゃろ？",
 		"。": "じゃ。",
-		"": "じゃ。"
+		"EOS": "じゃ。"
 	},
 	"まし": {
 		"た,か,？": "たかや？",
+		"た,ね,？": "たかのぉ？",
+		"た,か": "たかや",
+		"た,ね": "たかのぉ",
 		"た,。": "たでありんす。",
 		"た,？": "たかや？",
-		"た": "たでありんす。"
+		"た": "たでありんす"
 	},
 	"でし": {
 		"た,か,？": "だったかや？",
-		"た,ね,？": "だったかや？",
-		"た,ね": "だったかや？",
+		"た,ね,？": "だったかのぉ？",
+		"た,か": "だったかや",
+		"た,ね": "だったかのぉ",
 		"た,。": "だったでありんす。",
 		"た,？": "だったかや？",
-		"た": "だったでありんす。"
+		"た": "だったでありんす"
 	},
 	"ます": {
 		"か,ね": "かや？",
 		"よ,ね": "んじゃろ",
 		"か,？": "かや？",
 		"ね,？": "かや？",
+		"か,。": "かや",
+		"ね,。": "んじゃ。",
+		"わ,。": "んじゃ。",
+		"よ,。": "んじゃ。",
 		"か": "かや？",
-		"ね": "んじゃ。",
-		"わ": "んじゃ。",
-		"よ": "んじゃ。",
+		"ね": "んじゃ",
+		"わ": "んじゃ",
+		"よ": "んじゃ",
 		"。": "んじゃ。",
-		"": "んじゃ。"
+		"EOS": "んじゃ。"
 	},
 	"です": {
 		"か,ね,？": "かのぉ？",
 		"よ,ね,。": "じゃな。",
 		"か,ね": "かのぉ？",
-		"よ,ね": "じゃな。",
+		"よ,ね": "じゃな",
 		"か,？": "かや？",
 		"ね,？": "かや？",
+		"か,。": "かや",
 		"ね,。": "じゃ。",
 		"わ,。": "じゃ。",
 		"よ,。": "じゃよ。",
 		"か": "かや？",
-		"ね": "じゃ。",
-		"わ": "じゃ。",
-		"よ": "じゃよ。",
+		"ね": "じゃ",
+		"わ": "じゃ",
+		"よ": "じゃよ",
 		"。": "じゃ。",
-		"": "じゃ"
+		"EOS": "じゃ"
 	},
 	"だ": {
 		"よ,ね,。": "じゃな。",
@@ -85,10 +94,24 @@
 		"わ,。": "じゃ。",
 		"よ,。": "じゃよ。",
 		"けど": "じゃが",
-		"ね": "じゃ。",
-		"わ": "じゃ。",
-		"よ": "じゃよ。",
+		"ね": "じゃ",
+		"わ": "じゃ",
+		"よ": "じゃよ",
 		"。": "じゃ。",
-		"": "じゃ"
+		"EOS": "じゃ"
+	},
+	"てらっしゃい": {
+		"よ,ね,。": "じゃな。",
+		"よ,ね": "じゃな。",
+		"ね,？": "かや？",
+		"ね,。": "じゃ。",
+		"わ,。": "じゃ。",
+		"よ,。": "じゃよ。",
+		"けど": "じゃが",
+		"ね": "じゃ",
+		"わ": "じゃ",
+		"よ": "じゃよ",
+		"。": "じゃ。",
+		"EOS": "じゃ"
 	}
 }
