+++ models/holo/conv_honorific.json
@@ -2,10 +2,18 @@
 	"いらっしゃいます": "います",
 	"なさいますか": "しますか",
 	"なされますか": "しますか",
-	"なさいます": "する",
-	"なされます": "する",
+	"なさいます": "します",
+	"なされます": "します",
 	"されますか": "しますか",
 	"おられます": "います",
+	"おります": "います",
+	"おはようございます": "おはようさん",
+	"ござる": "ある",
+	"ござい": "あり",
+	"ござれ": "あれ",
+	"なさる": "する",
+	"なさい": "し",
+	"なされ": "し",
 	"おっしゃら": "言わ",
 	"おっしゃり": "言い",
 	"おっしゃる": "言う",
@@ -16,7 +24,8 @@
 	"こちら": "こっち",
 	"あちら": "あっち",
 	"されて": "して",
-	"ご覧にな": "見",
+	"ご覧になる": "見る",
+	"ご覧になって": "見て",
 	"ご覧": "見て",
 	"お話しください": "話してくりゃれ",
 	"お話し下さい": "話してくりゃれ",
