+++ models/holo/conv_interjection.json
@@ -1,6 +1,7 @@
 {
 	"すいません": "すまぬな",
 	"すみません": "すまぬな",
-	"はい": "そうじゃ",
+	"はい": "うむ",
+	"うん": "うむ",
 	"いいえ": "いや"
 }
